Decompose ActiveGroupSwitcher.tsx into smaller files
Split the 189-line switcher into three focused files per review feedback:
scopeDisplay.ts (pure label/icon/key helpers), ScopeMenuBody.tsx (the
dropdown list), and the trimmed-down ActiveGroupSwitcher.tsx (trigger +
back-to-default pill).
